melanoma/bot/requirements.txt
torch
numpy
pandas
pillow
torchvision
python-telegram-bot
efficientnet-pytorch
